+++ networkdriver/portmapper/mapper.go
@@ -4,7 +4,7 @@ import (
 	"errors"
 	"fmt"
 	"github.com/dotcloud/docker/pkg/iptables"
-	"github.com/dotcloud/docker/proxy"
+	"github.com/dotcloud/docker/pkg/proxy"
 	"net"
 	"sync"
 )
