+++ pkg/proxy/network_proxy_test.go
