+++ pkg/proxy/stub_proxy.go
