+++ utils/utils.go
@@ -879,16 +879,6 @@ func ShellQuoteArguments(args []string) string {
 	return buf.String()
 }
 
-func IsClosedError(err error) bool {
-	/* This comparison is ugly, but unfortunately, net.go doesn't export errClosing.
-	 * See:
-	 * http://golang.org/src/pkg/net/net.go
-	 * https://code.google.com/p/go/issues/detail?id=4337
-	 * https://groups.google.com/forum/#!msg/golang-nuts/0_aaCvBmOcM/SptmDyX1XJMJ
-	 */
-	return strings.HasSuffix(err.Error(), "use of closed network connection")
-}
-
 func PartParser(template, data string) (map[string]string, error) {
 	// ip:public:private
 	var (
